Переход на однофайловые vue-компоненты
Переход от компонент в .js файлах на однофайловые .vue компоненты.
Внесены изменения в настройки webpack для сборки vue-компонент.
