+++ 02 - students/Александр Прокофьев/project/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "dependencies": {
     "express": "^4.17.1",
-    "moment": "^2.24.0"
+    "moment": "^2.24.0",
+    "vue": "^2.6.10"
   },
   "devDependencies": {
     "@babel/core": "^7.6.4",
@@ -17,6 +18,8 @@
     "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.8.0",
     "nodemon": "^1.19.4",
+    "vue-loader": "^15.7.1",
+    "vue-template-compiler": "^2.6.10",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.9",
     "webpack-node-externals": "^1.7.2"
@@ -27,7 +30,7 @@
     ]
   },
   "scripts": {
-    "serve": "nodemon server/server.js",
+    "serve": "nodemon src/server/server.js",
     "build": "webpack --mode development --config webpack.config.js &&  webpack --mode development --config server.webpack.config.js",
     "serveprod": "node dist/server/server.js"
   },
