+++ 02 - students/Александр Прокофьев/project/src/public/index.js
@@ -5,6 +5,9 @@ As of Babel 7.4.0, this package has been deprecated in favor of directly includi
 import '@babel/polyfill'
 import './css/normalize.css'
 import './css/style.css'
-import appObj from './js/main'
+import Vue from "vue"
+import app from './js/main.vue'
 
-let app = new Vue (appObj)
+new Vue({
+    render: h => h(app)
+}).$mount('#app')
