--- 02 - students/Александр Прокофьев/project/src/public/js/error-notifier.js
@@ -1,15 +1,0 @@
-let error = {
-    data(){
-        return{
-            err: ''
-        }
-    },
-    methods:{
-        flashErrMsg(err_msg){
-            this.err = err_msg
-        }
-    },
-    template:   `<div class=err-note v-if="err.length != 0"><h4>{{err}}</h4></div>`
-}
-
-export default error
