+++ 02 - students/Александр Прокофьев/project/webpack.config.js
@@ -1,6 +1,7 @@
 const path = require('path')
 const HtmlPlugin = require('html-webpack-plugin')
 const MiniCss = require('mini-css-extract-plugin')
+const VueLoaderPlugin = require("vue-loader/lib/plugin")
 
 module.exports = {
     entry: path.resolve(__dirname, 'src', 'public', 'index.js'),/*указать точку входа*/
@@ -12,6 +13,11 @@ module.exports = {
     target: 'web',/*указать тип сборки*/
     module: { //указание по обработке файлов для сборки
         rules: [
+            {
+                test: /\.vue$/,
+                exclude: /node_modules/,
+                loader: 'vue-loader'
+            },
             {
                 test: /\.js$/,
                 exclude: /node_modules/,
@@ -33,6 +39,7 @@ module.exports = {
             template: 'src/public/index.html', /*указание пути до файла шаблона */
             excludeChunks: ['server'],
             filename: 'index.html' /*результирующий файл */
-        })
+        }),
+        new VueLoaderPlugin()
     ]
 }
